Refactor IUser to abstract User, extend and better pom
