+++ aniwash/src/main/java/aniwash/Test.java
@@ -8,7 +8,7 @@ public class Test {
     public static void main( String[] args )
     {
         UserFactory userFactory = new UserFactory();
-        IUser customer = userFactory.createCustomer();
+        User customer = userFactory.createCustomer("x",1,"s","s","s","x");
         customer.setName("John");
         System.out.println(customer.getName());
